requirements.txt
accelerate==0.24.1
diffusers==0.23.1
jinja2==3.1.2
langchain==0.0.321
llama-cpp-python==0.2.11
peft==0.6.2
pillow==9.1.1
rich==13.7.0
transformers==4.35.2
xformers==0.0.22.post7
